Rename dim to rank
