+++ crates/onnx-ir/src/lib.rs
@@ -1,10 +1,10 @@
 mod coalesce;
-mod dim_inference;
 mod from_onnx;
 pub mod ir;
 mod node_remap;
 mod proto_conversion;
 mod protos;
+mod rank_inference;
 pub mod util;
 
 pub use from_onnx::convert_constant_value;
